// src/app/shared/models/consult-vehicle.ts
export interface ConsultVehicle {
  amount: number;
  brandDescription: string;
  capacity: number;
  carBodyDescription: string;
  chassisNumber: string;
  classNumber: number;
  commercialName: string;
  contactCommentary: string;
  contactEmail: string;
  contactPhone: string;
  contractType: string;
  currentCoverageAmount: string;
  cylinderCapacity: number;
  date: string;
  finalDateInsurance: string;
  fuelDescription: string;
  govermentCode: string;
  groupCode: number;
  initialDateInsurance: string;
  insuranceEndDate: string;
  insuranceStartDate: string;
  motorNumber: string;
  owedYears: number;
  ownerEmail: string;
  ownerId: number;
  ownerIdType: number;
  ownerName: string;
  payId: string;
  period: number;
  plateClass: string;
  plateNumber: string;
  previousCoverageAmount: string;
  responseCode: number;
  responseDescription: string;
  rtvIndicator: string;
  serialNumber: string;
  styleDescription: string;
  taxValue: number;
  vinNumber: string;
  yearModel: number;
}